shared-libs/pyproject.toml
[tool.poetry]
name = "cloudscoreastraai-shared-libs"
version = "0.1.0"
description = "Shared libraries for CloudScore ASTRA AI microservices"
authors = ["CloudScore Team"]
readme = "README.md"
packages = [{include = "shared"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
kafka-python-ng = "^2.2.2"
structlog = "^24.1.0"
python-json-logger = "^2.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pgvector = "^0.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
